requirements.txt
Django==4.1.5
django-compressor==4.1
django-extensions==3.2.1
django-htmx==1.13.0
django-simple-history==3.2.0
django-sms==0.6.0
djangorestframework==3.14.0
pytailwindcss==0.1.4
requests==2.28.1
pre-commit==2.21.0
django-environ==0.9.0